% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adj.corbootci}
\alias{adj.corbootci}
\title{Adjust quantiles of the correlation bootstrap confidence interval based
on sample size. Only necessary for Pearson correlation.}
\usage{
adj.corbootci(n, nboot)
}
\arguments{
\item{n}{Sample size. For the comparison of two independent correlations,
n is the sum of the two sample sizes.}

\item{nboot}{Number of bootstrap samples.}
}
\value{
Indices used to get the 2.5th and 97.5th quantiles of
the bootstrap distribution.
}
\description{
Adjustments are available for nboot = 599, so scaling is required for
different nboot values. Adjustments work only for alpha = 0.05.
}
