% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{corbsub}
\alias{corbsub}
\title{Compute correlation for bootstrap indices}
\usage{
corbsub(isub, x, y, corfun, ...)
}
\arguments{
\item{isub}{A vector of length n. A bootstrap sample from the sequence of integers}

\item{x, y}{Two vectors of observations of length n.}

\item{corfun}{A correlation function which returns a correlation value.}
}
\value{
A bootstrap correlation value.
}
\description{
This function is used by other functions when computing
}
