% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_simplified.R
\name{kendall}
\alias{kendall}
\title{Compute Kendall's tau}
\usage{
kendall(x, y)
}
\arguments{
\item{x, y}{Two vectors of the same length.}
}
\value{
The correlation coefficient.
}
\description{
Simplified version of \code{kendall.test} which returns only the correlation coefficient.
Called by the bootstrap functions for faster execution.
}
