% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{kendall.test}
\alias{kendall.test}
\title{Compute Kendall's tau}
\usage{
kendall.test(x, y, alternative = "two.sided", alpha = 0.05)
}
\arguments{
\item{x, y}{Two vectors of the same length.}

\item{alternative}{Type of test, either "two.sided", "greater" for positive
correlations, or "less" for negative correlations.}

\item{alpha}{Used to compute 1-alpha confidence interval - default to 0.05.}
}
\value{
\itemize{
\item \code{estimate} the correlation coefficient.
\item \code{ci} the confidence interval for tau.
\item \code{p.value} the p-value of the test.
\item \code{statistic} test statistic used to compute the p value.}
}
\description{
Missing values are automatically removed.
Also compute 1-alpha confidence interval using
the method recommended by Long and Cliff (1997).
}
