% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mkcord}
\alias{mkcord}
\title{Make standard normal correlated data from 2 groups/conditions}
\usage{
mkcord(rho = 0, n = 50)
}
\arguments{
\item{rho}{the population correlation the sample is drawn from (default 0).}

\item{n}{the sample size.}
}
\value{
A list with elements x and y.
}
\description{
Make standard normal correlated data from 2 groups/conditions
}
