% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mkcord3}
\alias{mkcord3}
\title{Make standard normal correlated data from 3 groups/conditions}
\usage{
mkcord3(rho.x1y = 0, rho.x2y = 0.5, rho.x1x2 = 0.2, n = 50)
}
\arguments{
\item{rho.x1y}{the population correlation between x1 and y (default 0).}

\item{rho.x2y}{the population correlation between x2 and y (default 0.5).}

\item{rho.x1x2}{the population correlation between x1 and x2 (default 0.2).}

\item{n}{the sample size.}
}
\value{
A list with elements x1, x2 and y.
}
\description{
Make standard normal correlated data from 3 groups/conditions
}
