% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mkcord4}
\alias{mkcord4}
\title{Make standard normal correlated data from 4 groups/conditions}
\usage{
mkcord4(
  rho.x1y1 = 0,
  rho.x2y2 = 0.5,
  rho.x1x2 = 0.2,
  rho.y1y2 = 0.2,
  rho.x1y2 = 0,
  rho.x2y1 = 0,
  n = 50
)
}
\arguments{
\item{rho.x1y1}{the population correlation between x1 and y1 (default 0).}

\item{rho.x2y2}{the population correlation between x2 and y2 (default 0.5).}

\item{rho.x1x2, rho.y1y2, rho.x1y2, rho.x2y1}{the other population correlations.}

\item{n}{the sample size.}
}
\value{
A list with elements x1, y1, x2 and y2.
}
\description{
Make standard normal correlated data from 4 groups/conditions
}
