% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{pbcor.test}
\alias{pbcor.test}
\title{Compute the percentage bend correlation between x and y.}
\usage{
pbcor.test(x, y, beta = 0.2, alternative = "two.sided")
}
\arguments{
\item{x, y}{Two vectors of the same length.}

\item{beta}{The bending constant (default 0.2).}
}
\value{
\itemize{
  \item \code{estimate} the correlation coefficient.
  \item \code{p.value} the p-value of the test.
  \item \code{statistic} the t statistic of the test.
  }
}
\description{
Compute the percentage bend correlation between x and y.
}
