% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{spearman.test}
\alias{spearman.test}
\title{Compute Spearman's rho}
\usage{
spearman.test(x, y, alternative = "two.sided")
}
\arguments{
\item{x, y}{Two vectors of the same length.}

\item{alternative}{Type of test, either "two.sided", "greater" for positive
correlations, or "less" for negative correlations.}
}
\value{
\itemize{ \item \code{estimate} the correlation coefficient. \item
\code{p.value} the p-value of the test. \item \code{statistic} the t
statistic of the test. }
}
\description{
Missing values are automatically removed.
}
