% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{twocorci.nov}
\alias{twocorci.nov}
\title{Compare two dependent correlations: non-overlapping case}
\usage{
twocorci.nov(
  x1,
  y1,
  x2,
  y2,
  method = "pbcor",
  nboot = 2000,
  alpha = 0.05,
  alternative = "two.sided",
  null.value = 0,
  ...
)
}
\arguments{
\item{x1, y1, x2, y2}{Four dependent vectors of the same length.}

\item{method}{A function that returns a correlation coefficient.
Options include "spearman", "pbcor", "wincor". Default is "pbcor".}

\item{nboot}{Number of bootstrap samples. Default 2000.}

\item{alpha}{Alpha level. Default 0.05. For corfun = pearson, alpha is restricted
to 0.05 because the confidence interval adjustments have not been calculated for
other alphas.}

\item{alternative}{Type of test, either "two.sided" (default), "greater" for
positive correlations, or "less" for negative correlations.}

\item{null.value}{Hypothesis to test. Default 0.}

\item{...}{Optional parameter to pass to correlation function.}
}
\value{
\itemize{
\item \code{estimate.x1y1} the correlation coefficient between x1 and y1.
\item \code{estimate.x2y2} the correlation coefficient between x2 and y2.
\item \code{difference} the difference between correlation coefficients.
\item \code{conf.int} the bootstrap confidence interval.
\item \code{p.value} the bootstrap p value.
}
}
\description{
Compute a 1-alpha percentile bootstrap confidence interval for the difference
between two correlation coefficients corresponding to two dependent groups,
in the non-overlapping case. Compare correlation between x1 and y1 to the correlation between x2 and y2.
The default correlation is the percentage bend correlation.
This function is inappropriate to make inferences about Pearson's correlations.
Missing values are automatically removed.
}
\section{Note}{
 Modified from function \code{twoDNOV}
from Rallfun-v37.txt - see \url{https://github.com/nicebread/WRS/} and
\url{http://dornsife.usc.edu/labs/rwilcox/software/}.
}

\references{

}
