% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{twocorci.ov.sub}
\alias{twocorci.ov.sub}
\title{Compute correlation for bootstrap indices}
\usage{
twocorci.ov.sub(isub, x1, x2, y, corfun = pbcor, ...)
}
\arguments{
\item{isub}{A vector of length n. A bootstrap sample from the sequence of integers}

\item{x1, x2, y}{Three vectors of dependent observations of length n.}

\item{corfun}{A correlation function which returns a correlation value.}
}
\value{
A bootstrap correlation value.
}
\description{
This function is used by \code{twocorci.ov} when computing
}
