% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{wincor.test}
\alias{wincor.test}
\title{Compute the Winsorized correlation between x and y.}
\usage{
wincor.test(x, y, tr = 0.2, alternative = "two.sided")
}
\arguments{
\item{x, y}{Two vectors of the same length.}

\item{tr}{The amount of Winsorization (default 0.2).}
}
\value{
\itemize{
  \item \code{estimate} the correlation coefficient.
  \item \code{p.value} the p-value of the test.
  \item \code{statistic} the t statistic of the test.
  }
}
\description{
Compute the Winsorized correlation between x and y.
}
