% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{winval}
\alias{winval}
\title{Winsorize the data in the vector x.}
\usage{
winval(x, tr = 0.2)
}
\arguments{
\item{x}{A vector.}

\item{tr}{The amount of Winsorization, between 0 and 1 (default 0.2).}
}
\description{
Called by \code{wincor}.
}
